// src/app/components/tables/orders/types/types.ts
export type Orders = {
  id?: number | undefined | null;
  address: string | undefined | null;
  customer_name: string | undefined | null;
  email: string | undefined | null;
  order_date: string | undefined | null;
  order_number: string | undefined | null;
  payment_date: string | undefined | null;
  payment_status: number | undefined | null;
  payment_type: number | undefined | null;
  phone_number: string | undefined | null;
  shipment: number | undefined | null;
  status_order: string | undefined | null;
  total_harga: number | undefined | null;
  total_order: number | undefined | null;
};